requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
base58==2.1.1
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.3.0
bm25s==0.2.1
Brotli==1.1.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
chonkie==1.0.10
click==8.2.1
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
datasets==4.0.0
demjson3==3.0.6
dill==0.3.8
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
-e git+https://github.com/RUC-NLPIR/FlashRAG.git@e8ebc233af93c85b3693191db49f488077681faf#egg=flashrag_dev&subdirectory=../../../third_party/flashrag
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fschat==0.2.36
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.43.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jtutils==0.0.8
kiwisolver==1.4.9
langcodes==3.5.0
langid==1.1.6
language_data==1.3.0
lark-parser==0.12.0
latex2mathml==3.78.0
Levenshtein==0.27.1
llvmlite==0.44.0
lxml==6.0.1
marisa-trie==1.3.0
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib==3.10.5
mcp==1.13.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==2.1.2
networkx==3.5
nh3==0.3.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
openai==1.101.0
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
peft==0.17.1
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyjnius==1.6.1
pynndescent==0.5.13
pyparsing==3.2.3
pyserini==1.2.0
PySocks==1.7.1
PyStemmer==3.0.0
python-csv==0.0.14
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
-e git+ssh://git@github.com/tiagoriely/ai_squash_session.git@56217efc82ea85d135fcccd1cabf5a0b7bdc0a9f#egg=rag_squash_pipeline
rank-bm25==0.2.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.27.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==3.0.2
starlette==0.47.3
streamlit==1.48.1
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.4
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xlrd==2.0.2
xmltodict==0.14.2
xxhash==3.5.0
yamllint==1.37.1
yarl==1.20.1
